pyproject.toml
[tool.poetry]
name = "bfv-hacker-checker"
version = "0.2.1"
description = ""
authors = ["Albrecht Nitsche <albrechtnitsche@mail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
poethepoet = "^0.19.0"
pillow = "^9.4.0"
pyqtwebengine = "^5.15.6"
pynput = "^1.7.6"
numpy = "^1.24.2"
pytesseract = "^0.3.10"
injector = "^0.20.1"
humanfriendly = "^10.0"
pyqt5 = "^5.15.9"
pyqt5-tools = "^5.15.9.3.3"
pyqtdarktheme = "^2.1.0"
pywin32 = { version = "^306", platform = "win32" }
ewmh = { version = "^0.1.6", platform = "linux" }
opencv-python-headless = "^4.7.0.72"

[tool.poetry.dev-dependencies]
pyinstaller = "^5.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
app = "python src/main.py"
app-linux = "python src/main.py -c config.linux.ini"
build = "python scripts/build.py"
set-version = "python scripts/set_version.py"
gen-ui = ["gen-ui-app-window", "gen-ui-debug-window"]
gen-ui-app-window = "pyuic5 ui/app_window.ui -o src/ui_generated/Ui_AppWindow.py"
gen-ui-debug-window = "pyuic5 ui/debug_window.ui -o src/ui_generated/Ui_DebugWindow.py"



